mcp-server/pyproject.toml
[project]
name = "mcp-server"
version = "0.2.0"
description = "Multi-person workflow MCP server (FastMCP + Cosmos DB)"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.10.0",
    "azure-cosmos>=4.7.0",
    "azure-identity>=1.19.0",
    "azure-monitor-opentelemetry>=1.6.4",
    "pydantic>=2.9.0",
    "starlette>=0.40.0",
    "uvicorn>=0.32.0",
    "aiohttp>=3.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_server"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["src"]
testpaths = ["tests"]
